Sort AEP items by `id` instead of `order` to fix unsorted display issue
AEPs were being sorted by the `order` field, but since the majority of items lack this property, it caused them to appear unsorted. Changed the sort logic to use the `id` field instead, ensuring a consistent order. This can be revisited in the future if custom ordering is needed, but it will require the `order` field to be populated for all AEPs.